refactor: Toast 코드 리팩토링 (#9)
