+++ ensolution-front/src/common/components/index.ts
@@ -2,3 +2,4 @@ export * from './DetailPageHeader';
 export * from './FullPageLoader';
 export * from './EmptyState';
 export * from './Tabs';
+export * from './ToastProvider';
