+++ ensolution-front/src/common/contexts/toast.context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+import type { ToastType } from '@/common/ui/Toast';
+
+export interface ToastContextType {
+  showToast: (message: string, type?: ToastType) => void;
+}
+
+export const ToastContext = createContext<ToastContextType | undefined>(undefined);
